Rename existing AlertLayer to AlertLayerOld, prepare AlertLayer skeleton
