+++ scwx-qt/source/scwx/qt/map/alert_layer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <scwx/awips/phenomenon.hpp>
-#include <scwx/qt/map/map_context.hpp>
+#include <scwx/qt/map/draw_layer.hpp>
 
 #include <memory>
 #include <string>
@@ -14,19 +14,30 @@ namespace qt
 namespace map
 {
 
-class AlertLayerImpl;
-
-class AlertLayer
+class AlertLayer : public DrawLayer
 {
+   Q_DISABLE_COPY_MOVE(AlertLayer)
+
 public:
-   explicit AlertLayer(std::shared_ptr<MapContext> context);
+   explicit AlertLayer(std::shared_ptr<MapContext> context,
+                       scwx::awips::Phenomenon     phenomenon);
    ~AlertLayer();
 
-   std::vector<std::string> AddLayers(awips::Phenomenon  phenomenon,
-                                      const std::string& before = {});
+   void Initialize() override final;
+   void Render(const QMapLibre::CustomLayerRenderParameters&) override final;
+   void Deinitialize() override final;
+
+   bool RunMousePicking(
+      const QMapLibre::CustomLayerRenderParameters& params,
+      const QPointF&                                mouseLocalPos,
+      const QPointF&                                mouseGlobalPos,
+      const glm::vec2&                              mouseCoords,
+      const common::Coordinate&                     mouseGeoCoords,
+      std::shared_ptr<types::EventHandler>& eventHandler) override final;
 
 private:
-   std::unique_ptr<AlertLayerImpl> p;
+   class Impl;
+   std::unique_ptr<Impl> p;
 };
 
 } // namespace map
